Process files only once
Adds a state file in the config dir to record which files have already been processed.

That way if files are already in the watch folder, then a new file is added, we don't re-process the previously existing
files. (Deleting all our work from before by overwriting the files in the output dir!)

I hash the inode+size+filename and store that, for a little privacy.